Consolidate different versions of EmrOrderService
