+++ api/src/main/resources/moduleApplicationContext.xml
@@ -246,6 +246,27 @@
         <constructor-arg name="emrOrderService" ref="emrOrderService"/>
     </bean>
 
+    <bean id="emrOrderService" class="org.springframework.transaction.interceptor.TransactionProxyFactoryBean">
+        <property name="transactionManager">
+            <ref bean="transactionManager"/>
+        </property>
+        <property name="target" ref="emrOrderServiceTarget"/>
+        <property name="preInterceptors">
+            <ref bean="serviceInterceptors"/>
+        </property>
+        <property name="transactionAttributeSource">
+            <ref bean="transactionAttributeSource"/>
+        </property>
+    </bean>
+
+    <bean name="emrOrderServiceTarget" class="org.openmrs.module.emrapi.encounter.EmrOrderServiceImpl">
+        <constructor-arg name="openMRSDrugOrderMapper" ref="openMRSDrugOrderMapper"/>
+        <constructor-arg name="encounterService" ref="encounterService"/>
+        <constructor-arg name="openMRSOrderMapper" ref="openMRSOrderMapper"/>
+        <constructor-arg name="orderSetService" ref="orderSetService"/>
+        <constructor-arg name="openMRSOrderGroupMapper" ref="openMRSOrderGroupMapper"/>
+    </bean>
+
     <bean id="encounterObservationServiceHelper" class="org.openmrs.module.emrapi.encounter.EncounterObservationServiceHelper">
         <constructor-arg name="conceptService" ref="conceptService"/>
         <constructor-arg name="emrApiProperties" ref="emrApiProperties"/>
